+++ src/lib.rs
@@ -85,7 +85,7 @@ gmp-mpfr-sys in your crate, add it as a dependency inside
 gmp-mpfr-sys = "1.2"
 ```
 
-This crate required rustc version 1.31.0 or later.
+This crate required rustc version 1.37.0 or later.
 
 If the C libraries have a major version bump with some deprecated
 functions removed, but no features are removed in the Rust bindings,